moved misc chips to own file
